Hey man,
I've managed to add the sword to a socket on the character skeleton and I've got the attack animation working on click.
Also re adjusted the Menu and built lighting.
Just your stuff to do.

What I'd like done in order of priority is as follows
- Have the games tasks function (currently can't kill the rat).
- Have the days appear as they are, (Currently stops at day 2)
- Have the sword attach to the characters hand after they pick up the sword.
- Fix Rat movement.
- Fix any other bugs you may find.

I'll be available all of tomorrow for testing, you're even welcome to come to my studio if you so wish as well.